From Philippe and Axel: prevent deadlock.
A python thread has the GIL and is waiting for the ROOT lock.
Another Python thread has the ROOT lock and needs to print a Warning().
The latter will evoke a python call, waits for the GIL - deadlock.

As we cannot determine whether the Warning() thread has the ROOT lock
and as we can end up in this situation in a myriad ways, simply do not
risk requiring the GIL in ROOT-multi-threaded mode, by not passing
through the python message handler but through the default one of ROOT.